Refactor type declarations, separate concerns for better readability
This commit centralizes type declarations by creating 'enums.ts' and 'interfaces.ts' files for defining types in a single place. This will improve code readability and maintainability by separating concerns, as now type definitions are not scattered and tightly coupled with the logic in 'data.ts'.
Also, changes are made in "app.service.ts" to incorporate the types from the new files and service methods now take an id argument, enhancing functionality and specificity.
Next, this commit also improves the robustness of endpoint methods in "app.controller.ts" to throw an error if no report is found.
Overall, this commit helps make the code cleaner, more organized and consistent, and easier to understand for future enhancements.