Refactored update and delete methods for reports
The income and expense report update and delete methods were refactored into a single updateReport and deleteReport method to reduce code redundancy and improve maintainability.

An IUpdatedReport interface was also added, allowing optional inputs for report updates. The change will improve flexibility for report updates as not all fields need to be supplied for each update, only the ones that require a change.

The error messages for NotFound exceptions were also updated to include the provided id for better debugging.